[REFACTOR] Create package structure, move tests in package
