+++ pyprot/align/score.py
