+++ pyprot/structure/structure.py
